+++ CHANGELOG.md
@@ -5,5 +5,6 @@
     * SnooperAndroid.detailedPackageInfos
     * SnooperAndroid.sensorInfos
     * SnooperAndroid.systemFeatures
+    * SnooperAndroid.microphones
 * Added dependency on 'collections' pub package for deep-compare functionality
 * Addition of tests, example app & guidance documentation